chore: update tests
