+++ package.json
@@ -10,7 +10,7 @@
     "lint": "aegir lint",
     "build": "aegir build",
     "test": "aegir test",
-    "test:node": "aegir test --target node",
+    "test:node": "aegir test -t node -f \"./test/*.{node,spec}.js\"",
     "test:browser": "aegir test --target browser",
     "test:webworker": "aegir test --target webworker",
     "release": "aegir release --target node --target browser",
