+++ test/instance.node.js
@@ -2,33 +2,31 @@
 
 /* eslint-env mocha */
 
-const Client = require('..')
-const PeerId = require('peer-id')
 const multiaddr = require('multiaddr')
+const Stardust = require('../src')
 
-const IDJSON = require('./id.json')
-const SERVER_URL = multiaddr('/ip4/127.0.0.1/tcp/5892/ws/p2p-stardust')
-
+const { createPeer } = require('./utils')
 const mockUpgrader = {
   upgradeInbound: maConn => maConn,
   upgradeOutbound: maConn => maConn
 }
+const SERVER_URL = multiaddr('/ip4/127.0.0.1/tcp/5892/ws/p2p-stardust')
 
 describe('instance', () => {
   let client
   let conn
-  let id
+  let libp2p
 
   before(async () => {
-    id = await PeerId.createFromJSON(IDJSON)
+    [libp2p] = await createPeer()
   })
 
   it('should be creatable', () => {
-    client = new Client({ upgrader: mockUpgrader, id })
+    client = new Stardust({ upgrader: mockUpgrader, id: libp2p.peerInfo.id, libp2p })
     conn = client.createListener(() => {})
   })
 
   it('should connect to server', async () => {
-    await conn._listen(SERVER_URL)
+    await conn.listen(SERVER_URL)
   })
 })
